requirements.txt
asgiref==3.7.2
coverage==7.3.2
cssselect==1.2.0
Django==4.2.6
django-widget-tweaks==1.5.0
django-with-asserts==0.0.1
django-formtools==2.4.1
Faker==19.11.0
libgravatar==1.0.4
lxml==4.9.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
backports.zoneinfo;python_version<"3.9"